Started replacing Observable and Observer with Flow API. Applied code cleanup stuff.

+++ src/main/java/io/github/chr1sps/rars/riscv/instructions/AUIPC.java
@@ -51,6 +51,6 @@ public class AUIPC extends BasicInstruction {
     public void simulate(ProgramStatement statement) {
         int[] operands = statement.getOperands();
         RegisterFile.updateRegister(operands[0],
-                RegisterFile.getProgramCounter() - INSTRUCTION_LENGTH + (operands[1] << 12));
+                RegisterFile.getProgramCounter() - INSTRUCTION_LENGTH + ((long) operands[1] << 12));
     }
 }
