+++ src/main/java/io/github/chr1sps/rars/riscv/instructions/LB.java
@@ -46,6 +46,6 @@ public class LB extends Load {
      * {@inheritDoc}
      */
     public long load(int address) throws AddressErrorException {
-        return (Globals.memory.getByte(address) << 24) >> 24; // Shifting sign extends
+        return ((long) Globals.memory.getByte(address) << 24) >> 24; // Shifting sign extends
     }
 }
