+++ src/main/java/io/github/chr1sps/rars/riscv/instructions/LH.java
@@ -46,6 +46,6 @@ public class LH extends Load {
      * {@inheritDoc}
      */
     public long load(int address) throws AddressErrorException {
-        return (Globals.memory.getHalf(address) << 16) >> 16; // Shifting sign extends
+        return ((long) Globals.memory.getHalf(address) << 16) >> 16; // Shifting sign extends
     }
 }
