+++ src/main/java/io/github/chr1sps/rars/riscv/instructions/LUI.java
@@ -50,6 +50,6 @@ public class LUI extends BasicInstruction {
      */
     public void simulate(ProgramStatement statement) {
         int[] operands = statement.getOperands();
-        RegisterFile.updateRegister(operands[0], operands[1] << 12);
+        RegisterFile.updateRegister(operands[0], (long) operands[1] << 12);
     }
 }
